Fix 1.5.8 merge error
subdColors and meshColors are renamed to subdAndFriends and
meshAndFriends in 1.5.8 release

The fix to the reference import uses these two variables and they should
be renamed as well.

createSubD() and createPoly() here create a intermediate mesh object so
that its outMesh is connected to the target referenced mesh. Maya's
referenced mesh is read only. MFnMesh is not able to manipulate a
referenced mesh.