Fix importing per-vertex UVs
CL 981740

Polygon Winding is CW in Alembic and CCW in Maya. The code should
convert polygon winding when dealing with per-face-per-vertex UV
assignments. per-vertex UV assignment is not affected by polygon winding
since it's defined for each vertex..

per-face-per-vertex:
uvIds is in CCW and iSampIndices is in CW. So we need to convert polygon
winding.

per-vertex:
Both uvIds and vertexList are CCW. No need to convert winding.