EM race condition when evaluating fan-out connections in parallel
CL 994822

EM calls TevaluationGraph::makeDataSharable() method to make all Tdata
to be holders before evaluation. This makes the output
(AbcImport.outPolyMesh[0] plug) to be a holder. AbcImport just sets the
datablock clean in its compute() because there are no changes to the
mesh data. So the output remains to be a holder. Then EM evaluates the
fan-out destination mesh shapes in parallel. In TmeshShape's
computeObject(), they evaluate the inMesh connection and call
AbcImport.outPolyMesh[0] Tdata's makeWritable in parallel.. This breaks
reference counting and caused the mesh data to be deleted unexpectedly..